+++ rbx_types/CHANGELOG.md
@@ -1,6 +1,8 @@
 # rbx_types Changelog
 
 ## Unreleased Changes
+
+## 0.3.0 (2021-02-16)
 * Renamed `EnumValue` to `Enum`.
 * Added `Display` and `FromStr` implementations for `Ref`.